# ImageEditorAPI/openAPI/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/api/v1","description":"Generated server url"}],"tags":[{"name":"Image Filters Controller","description":"Базовый CRUD API для работы с пользовательскими запросами на редактирование картинок"},{"name":"Image Controller","description":"Базовый CRUD API для работы с картинками"},{"name":"Auth Controller","description":"Контроллер для регистрации/авторизации пользователя"}],"paths":{"/image":{"post":{"tags":["Image Controller"],"summary":"Загрузка нового изображения в систему","description":"В рамках данного метода необходимо:\n1. Провалидировать файл. Максимальный размер файла - 10Мб, поддерживаемые расширения - png, jpeg.\n1. Загрузить файл в S3 хранилище.\n1. Сохранить в БД мета-данные файла - название; размер; ИД файла в S3; ИД пользователя, которому файл принадлежит.\n","operationId":"uploadImage","requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"400":{"description":"Файл не прошел валидацию","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UiSuccessContainer"}}}},"200":{"description":"Успех выполнения операции","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadImageResponse"}}}},"500":{"description":"Непредвиденная ошибка","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UiSuccessContainer"}}}}}}},"/image/{image-id}/filters/apply":{"post":{"tags":["Image Filters Controller"],"summary":"Применение указанных фильтров к изображению","description":"В рамках данного метода необходимо:\n1. Проверить, есть ли у пользователя доступ к файлу\n1. Сохранить в БД новый запрос на изменение файла:\n    1. статус = WIP\n    2. ИД оригинальной картинки = ИД оригинального файла\n    3. ИД измененной картинки = null\n    4. ИД запроса = уникальный ИД запроса в системе\n1. Отправить в Kafka событие о создании запроса\n1. Убедиться, что шаг 3 выполнен успешно, в противном случае выполнить повторную попытку\n1. Вернуть пользователю ИД его запроса\n","operationId":"applyImageFilters","parameters":[{"name":"image-id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"filters","in":"query","required":true,"schema":{"type":"array","items":{"type":"string","enum":["REVERS_COLORS","CROP","REMOVE_BACKGROUND","OTHER"]}}}],"responses":{"404":{"description":"Файл не найден в системе или недоступен","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UiSuccessContainer"}}}},"500":{"description":"Непредвиденная ошибка","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UiSuccessContainer"}}}},"200":{"description":"Успех выполнения операции","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplyImageFiltersResponse"}}}}}}},"/auth/register":{"post":{"tags":["Auth Controller"],"summary":"Регистрирует нового пользователя","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/auth/login":{"post":{"tags":["Auth Controller"],"summary":"Авторизовывает зарегистрированного пользователя","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/images":{"get":{"tags":["Image Controller"],"summary":"Получение списка изображений, которые доступны пользователю","description":"В рамках данного метода необходимо:\n1. Получить мета-информацию о всех изображениях, которые доступны пользователю\n","operationId":"getImages","responses":{"500":{"description":"Непредвиденная ошибка","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UiSuccessContainer"}}}},"200":{"description":"Успех выполнения операции","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetImagesResponse"}}}}}}},"/image/{image-id}":{"get":{"tags":["Image Controller"],"summary":"Скачивание файла по ИД","description":"В рамках данного метода необходимо:\n1. Проверить, есть ли такой файл в системе.\n1. Проверить, доступен ли данный файл пользователю.\n1. Скачать файл.\n","operationId":"downloadImage","parameters":[{"name":"image-id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Успех выполнения операции","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}},"404":{"description":"Файл не найден в системе или недоступен","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UiSuccessContainer"}}}},"500":{"description":"Непредвиденная ошибка","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UiSuccessContainer"}}}}}},"delete":{"tags":["Image Controller"],"summary":"Удаление файла по ИД","description":"В рамках данного метода необходимо:\n1. Проверить, есть ли такой файл в системе.\n1. Проверить, доступен ли данный файл пользователю.\n1. Удалить файл.\n","operationId":"deleteImage","parameters":[{"name":"image-id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Файл не найден в системе или недоступен","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UiSuccessContainer"}}}},"500":{"description":"Непредвиденная ошибка","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UiSuccessContainer"}}}},"200":{"description":"Успех выполнения операции","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UiSuccessContainer"}}}}}}},"/image/{image-id}/filters/{request-id}":{"get":{"tags":["Image Filters Controller"],"summary":"Получение ИД измененного файла по ИД запроса","description":"В рамках данного метода необходимо найти и вернуть по ИД пользовательского запроса \nИД соответсвующего ему файла и статус, в котором находится процесс применения фильтров.\nПо ИД оригинального изображения нужно убедиться, что ИД запроса относится к нему и \nчто у пользователя есть доступ к данному изображению (оригинальному).\n","operationId":"getModifiedImageByRequestId","parameters":[{"name":"image-id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"request-id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Успех выполнения операции","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetModifiedImageByRequestIdResponse"}}}},"404":{"description":"Файл не найден в системе или недоступен","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UiSuccessContainer"}}}},"500":{"description":"Непредвиденная ошибка","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UiSuccessContainer"}}}}}}},"/auth/user":{"get":{"tags":["Auth Controller"],"summary":"Возвращает информацию о авторизованном пользователе","operationId":"user","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthUserDto"}}}}}}}},"components":{"schemas":{"UiSuccessContainer":{"required":["success"],"type":"object","properties":{"success":{"type":"boolean","description":"Признак успеха"},"message":{"type":"string","description":"Сообщение об ошибке"}}},"UploadImageResponse":{"required":["imageId"],"type":"object","properties":{"imageId":{"type":"string","description":"ИД файла","format":"uuid"}}},"ApplyImageFiltersResponse":{"required":["requestId"],"type":"object","properties":{"requestId":{"type":"string","description":"ИД запроса в системе","format":"uuid"}}},"RegisterRequestDto":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"maxLength":15,"minLength":5,"type":"string"}}},"LoginRequestDto":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"GetImagesResponse":{"required":["images"],"type":"object","properties":{"images":{"type":"array","items":{"$ref":"#/components/schemas/Image"}}}},"Image":{"required":["filename","size"],"type":"object","properties":{"imageId":{"type":"string","description":"ИД файла","format":"uuid"},"filename":{"type":"string","description":"Название изображения"},"size":{"type":"integer","description":"Размер файла в байтах","format":"int32"}}},"GetModifiedImageByRequestIdResponse":{"required":["imageId","status"],"type":"object","properties":{"imageId":{"type":"string","description":"ИД модифицированного или оригинального файла в случае отсутствия первого","format":"uuid"},"status":{"type":"string","description":"Статус обработки файла","enum":["WIP","DONE"]}}},"AuthUserDto":{"type":"object","properties":{"username":{"type":"string"},"roles":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}}}}}