# ImageEditorAPI/src/main/java/edu/tinkoff/imageeditor/dto/auth/AuthUserDto.java
package edu.tinkoff.imageeditor.dto.auth;

import com.fasterxml.jackson.annotation.JsonIgnore;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.util.Set;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class AuthUserDto {

    @JsonIgnore
    private String token;

    private String username;

    private Set<String> roles;

}